correct formatting of test suite
