+++ src/main/java/io/see/quick/JavadocGenerator.java
@@ -33,9 +33,6 @@ import java.io.IOException;
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 public class JavadocGenerator {
